requirements.txt
pywebview==5.2
pillow==10.4.0
